Separate library logic from UI components
React is an incredible tool for building UIs. However its power and
flexibility also make it easy to tightly couple the backend business
logic to the UI components.

This commit creates completely separate directory trees for "backend"
libraries and the front end components that display data to the user and
handle events/interactions.

The hope with this structure is that future developers can see the
directory structure as explaining what the code *does* instead of what
the code *is*.

For example, even though this app is made with React Native, I
deliberately chose the name `ui` for the React source files. This name
de-emphasizes the tool in favor of focusing on the value provided by the
directory's included code.

Similarly, `lib` wraps all the directories that I hope to open-source
into separate, standalone libraries for others to use in their own
projects. This design encourages me to think critically about the
subsystems that will compose this application, their inputs/outputs and
interactions with each other.