Create STIF Specification v0.1.0
This evening I found
[WCIF](https://github.com/SpeedcuberOSS/wcif/blob/master/specification.md#extension),
the official data exchange format used for official WCA competitions. I
reviewed it carefully to see if I could just use WCIF for the timer, but
the current answer is "no" because WCIF is highly tailored to the domain
of running a cubing competition, and doesn't provided `extensions`
points on `Attempt`s where I could mount in the extra data collected
during personal speedcubing practice.

As such, STIF, the Speedcubing Timer Interchange Format, has been born
with the purpose of providing a consistent set of fields for speedcubing
timers/analysis tools to implement for maximum inter-compatibility.