Add `ScrambleBuilder`
While adding the `ScrambleBuilder` I realized that the logic for
building/validating `id`s and `extensions` was identical to that of the
`AlgorithmBuilder`. Indeed, since all STIF types have `id` and
`extensions` fields it made sense to extract those out into an abstract
super type with an accompanying abstract super builder class (thus the
modifications to nearly all of the STIF specification files).

It took a while to figure out how to get TypeScript to correctly type
check the chained method calls since most of the examples for "Builder
Pattern with Inheritance" led to complex Java implementations full of
"Curiously Repeating Generic Patterns" (CSRP for short). Fortunately,
between this StackOverflow question
(https://stackoverflow.com/a/72679748/14765128), this Medium post
(https://levelup.gitconnected.com/typescript-advanced-types-this-type-and-dynamic-types-ecb99c4ec275),
and this section of the TypeScript documentation
(https://www.typescriptlang.org/docs/handbook/2/classes.html#this-types)
I was able to get it all figured out!