Organize Algorithm and Scramble Builder tests
The test suite adequately covered the expected feature set for both
builders, but it was difficult to read the tests to understand how best
to use the builders.

This commit re-organizes the test suite into a sort of "specification
format" with improved test names and groupings. This improved
organization should make it easier to read through the test suite to
understand how to use the Builders properly and to see the situations in
which the Builder will reject construction.

The organizing principles used in this commit were greatly inspired by
an NDC talk by Kevlin Henney about the Structure and Design of Unit
Tests that I found on YouTube