Create `InMemoryAttemptLibrary`
After completing a solve, we need to save attempts for later analysis.

For that purpose, this commit introduces the first iteration of an
`AttemptLibrary` which stores all the attempts in RAM (i.e. NOT
persisted). While certainly not the long-term solution, it provides a
solid baseline for designing the needed APIs without worrying too much
about the complexities of cross-platform persistent storage.