Enable `uuid` within React Native
`uuid` requires a function called `crypto.getRandomValues` which isn't
available by default in React Native.

We can "polyfill" it (i.e. substitute in an alternate implementation) by
importing the `react-native-get-random-values` package when the app
boots.

Thanks to https://github.com/LinusU/react-native-get-random-values for
making this fix so quick and easy to implement!