Improve `format_elapsed_time` function
`format_elapsed_time` is used to convert the current duration of an
attempt to a string to display on the screen.

The previous version had several problems, most notably the inability to
render times like `1:03:00.000` due to the failure to pad the `3`
seconds with zeros.

This new version comes with an extensive test suite that provides lots
of examples of the preferred string format for rendered times.