Apply Dark Theme by default
I personally prefer Dark Themes, so this change will make it more fun
for me to continue development.

Looking longer term, this commit functions as a proof of concept for
creating a custom theme, and loading it from behind an abstraction
boundary. This functionality will support multiple custom themes in the
future (a highly desired feature for speedcubing timers).