chore: Implement default story for AttemptCard
This serves as a proof of concept for using Storybook within this React
Native application.

Storybook currently sits on the "Learn" tab of the app. I will want to
move that to a deeper spot within a sidebar/settings as the app
development progresses (ultimately disabling it entirely in production
releases).