fix: Rewrite InspectionTimer
The old version was clunky and buggy. The library I had chosen to use
for the fancy circle timer animation forced me to tightly couple the
infraction logic to the display of the timer.

This commit makes a controlled component `InspectionTime` which is
entirely responsible for showing the currently remaining inspection time
(or the infraction status if applicable). Then a separate
`InspectionTimer` component runs the actual timer, tracks presses and
interactions, and reports the completion of inspection.