feat: Storybook Snapshot Testing
This is a huge step forward. I can now design my components visually,
but still get an automatic test suite to verify that the continue
rendering as expected across changes in the codebase.

This commit also includes automatic discovery of stories so that there's
no more need to continuously update the `index.js`.

I did have to move the `storybook` folder up into the root of the
project to get everything to work (not my favorite since it is purely a
UI tool), but I wanted to commit a working state before trying to move
it back into the `ui` folder.