fix: Switch Attempt.timestamp to a `number`
Attempt.timestamp was previous a Javascript Date object, which
serializes to a string, preventing an easy deserialization.

This commit switches the datatype to a number representing the
unixTimestamp of the attempt to facilitate the serialization.
Interestingly, this also simplified some downstream code that was
comparing the timestamps as part of a sorting algorithm.