fix: Enable persisting Attempts once again!
Fixing this issue was the cause of many hours of angst today.

For the last few weeks AsyncStorage simply hasn't worked, nothing could
be loaded from or saved to it. Inside of the debugger, I could execute
console.log statements right around the AsyncStorage calls, but the
async/await calls would seem to never execute.

I tried manually invoking the `await`ed command in the Debug Console but
always got stuck with a "Pending" promise. This was extremely
frustrating until I learned that such promises could not be resolved
while paused on a breakpoint
(https://stackoverflow.com/a/66925024/14765128)

I also tried to directly invoke commands against the `AsyncStorage`
object imported at the top of `AsyncStorageAttemptLibrary.ts` only to
discover that the Debug Console didn't know anything about such an
object. This was really confusing since I'm used to running `import x`
while working in Python and seeing the module appear in my local/global
variables list. Apparently that doesn't work in the JavaScript ecosystem
because the `import` statement is syntactic sugar for a complex
process carried out at "bundle-time" where all the separate JS/TS files
in my project are combined into a single monolithic bundle that the app
will load at boot. During that bundling process, the nice names like
`AsyncStorage` are converted to something else depending on the chosen
bundler (in the case of Metro Bundler, the default for React Native, it
looks like `AsyncStorage` turns into `_asyncStorage`).

In the end, the change that actually worked was disabling Storybook. It
seems like Storybook has some built-in polyfills/monkey-patches that are
creating destructive interference with my main application code.

I would like to keep Storybook easily available during development, but
it's obviously more important to have a functional application, so it's
getting sidelined for now.