fix: Enable connecting to a Rubik's Connected!
It seems like the issue caused by Storybook described in the last commit
was also the reason why the Bluetooth module was unable to carry out the
Bluetooth Connection to my Rubik's Connected.

In trying to debug the Bluetooth connection earlier today I had already
refactored the code to use the async/await paradigm over promises so I
kept it after things started working once Storybook was disabled.

We still aren't pulling data from the cube yet, but we are getting a
full connection and discovering the complete set of services for pulling
all the turns and other data sent from the cube.