feat: Create `BluetoothPuzzle` interface
This is the first step towards a platform-agnostic Bluetooth puzzle API
(i.e. compatible with web and mobile).

The key idea is to create a generic interface with all the methods
required for interacting with a BluetoothPuzzle, and then have unique
implementations of that interface on each platform.

In the case of this app, the `BluetoothPuzzle` interface is implemented
using `react-native-ble-plx`. When we're ready to merge into
`cubing.js`, then we'll create a new implementation of `BluetoothPuzzle`
based on the browser web bluetooth APIs.