feat: Create `RawGoCubeMessage`
To make it easier to work with the specific parts of the binary message
returned from the smartcube we will parse them into a specialized object
which provides named access to each part of the binary message.

Next steps are to convert the raw message into a more meaningful
`Rotation`, `Orientation`, and/or `BatteryLevel` type messages.