feat: Update UI right after discovering a smart puzzle.
The current Bluetooth scan is hard coded to take 5 seconds, after which
it will report back any smart puzzles it found.

This commit adds a callback to the Bluetooth Scan which will be called
whenever a smart puzzle is discovered. The UI passes in a callback which
updates the UI when needed.

closes #79