feat: POC Enable animating puzzles with cubing.js
Cubing.js is a web-based library, so it only runs well in a WebView.
However, WebViews can't easily communicate with the React Native engine,
so we have to create some workarounds to proxy messages back and forth.

In this commit, I added a hidden button to the WebView rendering the
cubing.js `twisty-player`. When clicked, that hidden button add its
current button text to the `twisty-player` as a new move.

The React Native pieces then post JS messages to change the button text,
then programmatically click it to add the move to the player.

I tried to directly invoke the `player` object, but apparently it's out
of scope for the `injectJavascript` calls. Thus the need for a reactive
workaround.