feat: Use `Rubik` font for timers
Adds the custom Rubik fonts to the app and applies them to the solve
timer.

The new look is beautiful! I'm thrilled to have finally made this
progress (I fought fonts for hours a month or two ago, ultimately giving
up and settling for the poorer look).

Next steps:
 - Add the fonts to the full theme so they apply throughout the whole UI
 - Change the timer fonts to Rubik Monospace to prevent jumping.

improves #85