feat: Check location/bluetooth permissions + status
Discovering Bluetooth devices requires the user to grant appropriate
permissions and enable related device services.

Specifially, on Android, the app needs to have Location/Bluetooth while
on iOS only Bluetooth is necessary.

This commit adds checks for all necessary permissions/services whenever
the user tries to discover smart puzzles. When possible, permissions are
requested as part of the flow.

Currently, if location/bluetooth is disabled, an error message is shown
requesting that the user enable the service. In the future, we should
seek to enable the service automatically with user approval.

NOTE: This commit starts the process of switching to a new Bluetooth
library. At this point, smart puzzles are no longer discoverable.