refactor: Remove `react-native-system-setting`
The package `react-native-system-setting` was the only one I found which
offered a method to check if a device's location services were enabled
(required for Android Bluetooth to work in API 23+). However, that
library maintainer also has a track record of choosing implementations
which violate Apple's policies, resulting in threats of termination of
Apple Developer Accounts. [1][2]

I don't need any of the other functionality in that library, and I
really don't want to expose myself to that author's implementation
whims. As such, I found it prudent to take this opportunity to learn how
to build a Native Module for React Native applications (at least for Android).

I largely followed the official guide [3], and the only hiccups I had
resulted from Android Studio selectively auto-saving files, causing
confusing, but temporary compilation errors.

[1] https://github.com/c19354837/react-native-system-setting/issues/28
[2] https://github.com/c19354837/react-native-system-setting/issues/58
[3] https://reactnative.dev/docs/native-modules-android