perf: Release `messages` memory after an Attempt.
Cubes with gyroscopes generate thousands of messages per minute. The
existing implementation stored all those messages in-memory throughout
the entire uptime of the application

This commit removes the `SmartPuzzleRegistry`'s `messages` repository.
If the full message history is required, it can be obtained by
registering the appropriate listener and saving it externally.

closes #88