chore: Update `react-native-paper` to `5.1.0`
This upgrade involved quite the bug chase. Upgrading the library
immediately broke all builds with a vauge error message of
"PhaseScriptExecution failed with a non-zero exit code".

Some research helped me realize that the failing script was part of the
React Native build process, but installing CocoaPods, resetting
node_modules, build caches, etc. still did not fix the issue.

The first insight came when re-installing node_modules. A new warning
had appeared saying that `react-native-paper` was missing a peer
dependency of `react-native-safe-area-context`. Installing the
dependency resolved the warning, but did not fix the build.

A key insight came after a restart of Metro Bundler. At the top of the
logs were two warnings, notifying of two packages which "[had] been
ignored because it contains invalid configuration". One of those
packages was `react-native-safe-area-context`. Upon closer inspection of
the error message, and cross referencing with the release notes I
realized that the "erroneous" configuration was actually targeted at
newer versions of React Native. The solution then became obvious:
downgrade the library to the version which targeted my version of React
Native.