fix: Assign specific `id` to `UNKNOWN` builtins
The `UNKNOWN` builtins serve as implementations of the Null Object
Pattern to simplify client code working with STIF objects.

It's important for clients to easily recognize an `UNKNOWN` object,
which wasn't previously possible since the `id` changed on each boot of
the app.

With this commit, every `UNKNOWN` entity has an `id` of `'unknown'` so
it can be easily recognized.

closes #91