fix(storybook): Enable easy booting for storybook
The command `yarn android:<environment>` now reboots the app to the
correct environment, including Storybook for React Native.

This should make it *much* easier to build a component library using
Storybook.

The "trick" here is I am now replacing the entire `index.js` file when
configuring a new environment. That way, when the bundler boots up it
only sees the dependencies required for the environment instead of all
dependencies for all environments controlled by a few boolean guards.

This also means that if the `index.js` needs to be updated for any
reason, those updates have to be mirrored across all environments.