fix(AttemptPlayer): Eliminate WebView re-renders
The WebView in TwistyPlayer re-renders every time the `algorithm` prop
changes. Those interruptions look like a complete fade to black followed
by a 0.5 second delay before re-drawing the cube with the updated
algorithm. These delays make it nearly impossible to appreciate a solve
replay.

This commit adds an imperative ref to the WebView so that the
AttemptPlayer can direct the TwistyPlayer WebView to update its rendered
alg, without re-rendering the entire WebView.

The result is a beautiful, slider-controlled replay of a solve :D