fix(android): Only execute `onScrubEnd` when scrubbing
The Android implementation of `react-native-slider` invokes the
`onSlidingComplete` callback every time the slider moves, instead of
only when a finger stops adjusting the slider like on iOS.

Since the `onScrubEnd` method invokes `setStartedPlayingAt`, it is not
guaranteed to keep time faithfully (i.e. playback may occur more slowly
than real-time). With Android calling `onScrubEnd` automatically, it was
resulting in an slower solve replay than expected.

Since Android still only invokes `onSlidingStart` when a finger touches
the slider, we can solve the problem by adding a guard over the logic of
`onScrubEnd` to keep it from executing unless scrubbing is occurring.