feat: Focus speed settings
Solve replays will care much more about slowing down the replay than
speeding it up, so there is not a need for the faster speeds.

On the flip side, too slow of a speed is not helpful either.

This commit focuses the set of speeds to those that will be most
commonly used. Others can be added in later if desired.