feat: Compress double turns
Smartcubes notify on every 90 degree rotation, but speedcubers commonly
execute 180 degree rotations which should be reported as such for proper
TPS statisitics.

A naive approach to recognizing a 180 degree turn is to find consecutive
moves with the same face and direction that occurred within a fixed,
small window.

However, this naive approach does not properly account for the varying
turn speeds of solvers at different levels of expertise. A 180 degree
turn takes far longer for a slow solver than for a fast solver. As such,
a dynamic window is required.

This commit adds a dynamic window computed as the average time per face
turn (total solve time / number of moves). When two identical moves
occur within that dynamic window they are considered a 180 degree turn.