feat: Include pickup/put down durations
The "pickup" phase is the period of time between starting the solve and
executing the first turn.

The "put down" phase is the period of time between solving the puzzle
and stopping the timer.

The adjustments required to compute these phases also made the time
durations for the other steps more accurate by including the
"recognition" time.