refactor!: Simplify STIF
The extensions feature from WCIF is neat, but prompted needlessly
deep data structures which complicate usage.

This commit re-writes the entire STIF specification to a much simpler
format which will be easier to work with. Interestingly, because this
rewrite makes several types more concrete, this update actually
strengthens validation of the data format and opens the door to further
strengthening in the future.