feat(stif): Support empty Reconstructions
The existing wrappers and builders for STIF expected every Solution to
have at least one SolutionPhase, and every SolutionPhase to have at
least one TimestampedMove. While that assumption makes the
implementation easier, it greatly complicates client code because it
breaks the contract of the corresponding STIF types (which allow empty
arrays).

This commit adds support for empty Reconstructions and SolutionPhases so
that client code can use the wrappers and builders with any valid STIF
data, even if it doesn't have any moves or SolutionPhases.