refactor: Rename `solveStart|End` to `timerStart|Stop`
The names `solveStart` and `solveEnd` make perfect sense in the context
of Attempts with only one puzzle/solution. However, for multi-puzzle
attempts, the term "solve" becomes ambiguous (Does 'solve' refer to the
entire attempt, or one solve within the attempt? If one solve, which one
out of many?).

In contrast, the terms `timerStart` and `timerStop` are much clearer --
obviously referring to the moments when the competitor starts and stops
the timer.

Furthermore, by referencing the timer in the name of the timestamp, it
makes it easier for newcomers to understand why the `duration` of an
attempt differs from the difference of `timerStop` and `timerStart` when
infractions (i.e. penalty time) have occurred.