chore: Remove Storybook
Storybook claims to be a wonderful tool for component driven
development, enabling isolated rendering and testing of all your UI
components.

Unfortunately, for me, that promise has not held up in this project.
While I did manage to configure Storybook and use it to develop a few
stories, it's caused far more pain and frustration than it has resolved.

Tonight I attempted to upgrade to Storybook for React Native 6.5, which
claims to solve many of the problems I'd been experiencing -- most
notably breaking Promises :/ Unfortunately, I was unable to render even
a simple "HelloWorld" story.

In total, Storybook has helped me with the development of a dozen or so
components, but has cost me at least 2 full working weeks worth of
development time in failed setup/configuration. I've given the tool
chance after chance and it consistently fails to deliver.

At this point, I'm going to take the route used by the
`react-native-paper` and `react-native-elements` teams: build a
collection of views which render key components with a variety of
formats as a sort of "kitchen sink". It may not have the fancy, dynamic
props editing from Storybook, but it also won't introduce any new
cryptic errors.

So, Storybook, this is the end of our story. The End.