refactor: Switch 3x3x3 Multi-BLD builtin to constant
3x3x3 Multi-BLD is tricky because the same event can have a different
number of puzzles. Until now, the STIF built-in for 333mbf has been a
function which first asks for the number of puzzles to solve before
building the resulting event.

However, since all other built-in events exactly match the
STIF.CompetitiveEvent interface, the unique typing of 333mbf greatly
increased the difficulty of building tools to iterate over all
built-ins.

This commit turns the built-in 333mbf event to a constant consisting of
two 3x3x3 puzzles. When more puzzles are desired, the builtin can be
spread into a new object with the desired number of puzzles.