feat: Improve SmartPuzzleCard usability
The Bluetooth connection status is now representated as a switch instead
of a button, which more clearly indicates what the status is and how to
change it.

The puzzle type icon has also been moved to the left of the card to make
it more prominent. This change also enabled better ellipsizing of extra
long puzzle/brand names.