refactor: Move tests for Examples to component folders
The snapshot file was huge (over 80k lines). Additionally, several
examples are not compatible with snapshot testing (e.g. interactivity,
timers, etc.) so they were unable to be used in the Examples explorer UI
since that would also trigger failing snapshot tests.

This commit removes the automatic snapshot testing for each component in
favor of allowing developers to opt-in to testing of their components.
The opt-in process is trivial: pass the example set to a `snapshots`
function in a test file. There's even a VS Code snippet for it!

Now snapshot files are much smaller and located closer to the components
they document.