fix(stif): Extract validation to break require cycle
The Wrappers were using Builders to validate the data they received, but
the Builders have an option to produce a Wrapped STIF entity -- creating
a cycle.

This commit extracts the validation logic so that Wrappers can use it
without needing to know that the Builders exist, breaking the require
cycle.