feat(persistence): Configure Realm database
While I'm not a fan of Realm's lock to Mongo Cloud for syncing, the
set-up for the open-source local version is far better than any other
database I've tried.

Within an hour or two (including distractions), I had my schema created
and was able to start saving/querying data.

Since the lion's share of the work on the app for the forseeable future
will be offline-only I won't need to worry about forking over tons of
money to Mongo. When I'm ready to start syncing though, it looks like it
will be a nearly trivial way to introduce a valuable "pro" (i.e. paid)
feature to the app so I can start earning some revenue for it.

Notice that nowhere in the UI components is there any mention of Realm.
That is a deliberate separation of concerns/duties since I want to put
all persistence code behind an abstraction boundary from the rest of the
application.

If I ever switch off of Realm in the future, I want to limit the blast
radius of changes.

closes #68