feat: Create a jsonl parser
JSONL is a nice format for streaming a list of JSON objects.

https://jsonlines.org/

This commit adds a parser which will work nicely with `read`ing
arbitrary length consecutive chunks of a JSONL file. Each detection of
a newline will emit the preceeding JSON value, whether that value was
part of a single read chunk, or required reading multiple chunks.

This parser does assume that the file ends with a newline character,
which is not guaranteed by the specification. Failure to include one
will result in the last item being omitted from the parse.