feat: Lazy load table averages + add paging
When the user has thousands of attempts, computing the averages can take
a while.

This commit allows the table to begin rendering almost immediately,
unblocking the thread while the averages are computed.

This is a stop-gap solution that works well enough for up to 5-10k
attempts. Long-term we'll need more performant averages computation
(e.g. a push model like I built in the jsonl parser -- add a value and
out pops a new average).