fix: Enable Bluetooth scans on Android 12+
Apparently, react-native-ble-plx version 2.0.3 was not updated to
support Android 12. While they have released version 3.0.0 which
claims to support Android 12, it is unstable -- crashing the app
for anything but a scan.

So, I switched to react-native-ble-manager which has had Android 12
support since 2021. I remember trying that library months ago and
feeling concerned about something, but that certainly wasn't the
case with this re-implementation. The re-write went smoothly and the
new library's API made it trivial to fix some other long-standing
bugs with BLE (i.e. unexpected puzzle disconnects).