feat(ticker): Add horizontal icons + value transforms
Adds a `transform` function prop to the `Ticker` component which can map
the integer `value` to any string or even custom JSX. This change
enables the Ticker to be used for stepping through an array of items
with custom rendering.

As a part of this change, added the option for horizontal arrow icons
for a more natural "scrolling" look.