feat: Enable assigning smartpuzzles to scrambles
In events with multiple puzzles, speedsolvers may want to use multiple
smartcubes to track every solution in the attempt.

This commit updates the `ScramblingView` component to support linking a
smartpuzzle to a specific scramble.

This update is momentous because it is the first time that any smartcube
application has taken the step towards concurrent move tracking from
multiple smartpuzzles. Truly a distinguishing feature of the app!