fix: See first scramble after changing events
Switching from index 9 of Multi BLD to a small relay resulted in a weird
scenario where the underlying Ticker would still be on index 9, even
though that was beyond its allowed bounds for the new event.

Now the ticker will reset to 0 when the event changes, which is the
expected behavior.