chore: Upgrade `react-native-ble-manager` to the New Architecture
This commit looks deceptively simple. It was one of the hardest
parts of the upgrade.

First, `LocationStatus.ts` was part of a custom native module I
had written to support the Android versions which require
location permissions for Bluetooth scans. Since that was on the
Old Architecture, I had to learn how to write a New Architecture
native module. That ended up being easiest to do by publishing
a separate package on npm.

Additionally, the maintainers of `react-native-ble-manager` are
still working on full support for the New Architecture. Timing was
very much in my favor since they started publishing viable release
candidates mere days before I started this upgrade project, and
the full release was published only ~20 hours ago.