chore: Upgrade `realm` to the New Architecture
MongoDB, the company that acquired Realm several years ago,
recently announced that they are deprecating Realm :/

So, I'll need to start building an alternate mass storage solution.
Fortunately, it looks like there's a new kid on the sqlite block,
`react-native-nitro-sqlite`, so I may be able to switch to the
database I had wanted to use in the first place.

BUT... that migration doesn't need to happen today. Realm still
works for now, so we'll keep it and upgrade it until we are able
to move to a more stable datastore.