chore: Upgrade to React Native 0.76
This has been an intensely difficult upgrade.

I attempted to use the upgrade guide, but ran into problem after
problem. The approach that ultimately proved successful was to
scaffold a brand new app, copy in my old code, and make
adjustments as needed to fix bugs.

This commit contains the core-only changes needed for the 0.76
upgrade. Subsequent commits will contain the third-party changes
required for full functionality on 0.76