chore: Upgrade `react-native-vector-icons` to New Architecture
The best part is that this upgrade **simplifies** my code and processes.

I no longer have to do the awkward manual linking steps in
Android and iOS. In fact, because `react-native-vector-icons` now
installs all fonts in `assets/fonts`, it also handles my custom fonts!

This commit also removes all the unused icon packs. My hope is
that will contribute to a smaller bundle size.

The one downside is that I have to add a bunch of mocks for my tests
to pass again. Hopefully that will be fixed more cleanly in the future.
 - See: https://github.com/oblador/react-native-vector-icons/issues/1681#issuecomment-2585560725